Added middelware to require Auth for protected routes. Avoid the anti-pattern of page and component code.
