--- app/auth/account/page.tsx
@@ -1,31 +1,0 @@
-import React from 'react';
-import { createServerClient } from '@supabase/ssr';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
-
-
-import Account from '../../../components/auth/account'
-
-export default async function LoginPage() {
-  const cookieStore = cookies()
-
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!, 
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!, 
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value
-        },
-      },
-    }
-  )
-  const { data } = await supabase.auth.getSession();
-
-  if (!data?.session) {
-    redirect('/login')
-  }
-
-  return <Account />
-
-}
