+++ app/auth/update-password/page.tsx
@@ -1,30 +1,25 @@
 
 import React from 'react';
-import { createServerClient } from '@supabase/ssr';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
+import NavBar from '@/components/NavBar';
 
+const UpdatePassword = () => {
+  return (
+    <div>
+      <NavBar />
+      <div className="flex justify-center items-center h-screen bg-zinc-950">
+        <div className="p-6 max-w-sm w-full bg-white rounded-lg border border-gray-200 shadow-md">
+          <h2 className="mb-4 text-xl font-bold text-gray-900">Update Password</h2>
+            <form>
+              <div className="mb-4">
+                <label htmlFor="password" className="block mb-2 text-sm font-medium text-gray-900">New Password</label>
+                <input type="password" id="password" className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5" required />
+              </div>
+              <button type="submit" className="w-full text-white bg-blue-600 hover:bg-blue-700 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center">Update Password</button>
+            </form>
+        </div>
+      </div>
+   </div>
+  );
+};
 
-import UpdatePassword from '../../../components/auth/update-password'
-
-export default async function UpdatePasswordPage() {
-  const cookieStore = cookies()
-  const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!, 
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!, 
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value
-        },
-      },
-    }
-  )
-  const { data } = await supabase.auth.getSession();
-
-  if (data?.session) {
-    redirect('/student')
-  }
-
-  return <UpdatePassword />
-}
+export default UpdatePassword;
