+++ app/layout.tsx
@@ -2,7 +2,6 @@
 
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import NavBar from '../components/NavBar'
 import './globals.css'
 
 const inter = Inter({ subsets: ['latin'] })
@@ -20,7 +19,6 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <NavBar />
         {children}
       </body>
     </html>
