+++ components/NavBar.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import Link from "next/link";
-import { createBrowserClient } from '@supabase/ssr';
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { FaBars, FaTimes } from "react-icons/fa";
 
 interface NavLink {
@@ -11,32 +10,13 @@ interface NavLink {
   title: string;
 }
 
-const Navbar = () => {
-  const supabase = createBrowserClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  )
+interface NavbarProps {
+  links?: NavLink[];
+}
 
+const Navbar = ({ links }: NavbarProps) => {
+  
   const [nav, setNav] = useState(false);  
-  const [loggedIn, setLoggedIn] = useState(false);
-
-  useEffect(() => {
-    const user = supabase.auth.getUser();
-    setLoggedIn(!!user);
-
-    supabase.auth.onAuthStateChange((event: any, session: any) => {
-      setLoggedIn(!!session?.user);
-    })
-
-  }, []);
-
-  const links: NavLink[] = loggedIn 
-  ? [
-    { id: 1, link: "auth/account", title: "Account" }
-  ]
-  : [
-    { id: 1, link: "auth/login", title: "Login" }
-  ]
 
   return (
     <div className="flex justify-between items-center w-full h-14 px-4 text-white bg-zinc-800 nav">
@@ -50,7 +30,7 @@ const Navbar = () => {
       </div>
 
       <ul className="hidden md:flex">
-        {links.map(({ id, link, title }) => (
+        {links && links.map(({ id, link, title }) => (
           <li
             key={id}
             className="nav-links px-4 cursor-pointer capitalize font-medium text-gray-500 hover:scale-105 hover:text-white duration-200 link-underline"
